+++ src/components/Header/Header.tsx
@@ -1,9 +1,10 @@
-import React, {ReactElement} from 'react';
+import React from 'react';
 import {View, StyleSheet} from 'react-native';
 
-import {InfoComponent} from './InfoComponent/InfoComponent';
-import {MainTaskButton} from './buttons/MainTaskButton';
-import {RankingButton} from './buttons/RankingButton';
+import {InfoComponent} from '../InfoComponent/InfoComponent';
+import {MainTaskButton} from '../buttons/MainTaskButton';
+import {RankingButton} from '../buttons/RankingButton/RankingButton';
+import {HeaderProps} from './type';
 
 const newParagraph = '\n\n\t';
 
@@ -12,12 +13,7 @@ const campaignDescription = `
     Sit amet dictum sit amet justo donec enim diam. Nisl pretium fusce id velit ut tortor pretium. Amet est placerat in egestas erat imperdiet sed euismod nisi. 
     Vitae nunc sed velit dignissim sodales ut eu sem. Adipiscing tristique risus nec feugiat. Pellentesque habitant morbi tristique senectus et netus et. Commodo ullamcorper a lacus vestibulum sed arcu non odio.\n\t Gravida neque convallis a cras semper auctor neque vitae. Cum sociis natoque penatibus et. Tincidunt lobortis feugiat vivamus at augue eget arcu dictum. Nisl vel pretium lectus quam id. Felis bibendum ut tristique et egestas\n\n\t quis ipsum suspendisse. Egestas dui id ornare arcu odio ut sem nulla. Quam vulputate dignissim suspendisse in est ante. Sed arcu non odio euismod lacinia at quis risus. Scelerisque viverra mauris in aliquam sem.`;
 
-type Props = {
-  exp: number;
-  navigate: (page: string) => void;
-};
-
-export const Header: React.FC<Props> = (props) => {
+export const Header: React.FC<HeaderProps> = (props) => {
   return (
     <View style={styles.container}>
       <RankingButton {...props} />
