+++ src/components/Header/type.ts
@@ -1,0 +1,4 @@
+export type HeaderProps = {
+  exp: number;
+  navigate: (page: string) => void;
+};
