+++ src/components/buttons/RankingButton/index.ts
@@ -1,0 +1,1 @@
+export {RankingButton} from './RankingButton';
