+++ src/components/buttons/RankingButton/type.ts
@@ -1,0 +1,4 @@
+export type RankingButtonProps = {
+  exp: number;
+  navigate: (page: string) => void;
+};
